+++ api/src/main/java/com/flipkart/poseidon/api/APIBuildable.java
@@ -16,18 +16,16 @@
 
 package com.flipkart.poseidon.api;
 
-import com.fasterxml.jackson.databind.JavaType;
 import com.flipkart.hydra.task.Task;
 import com.flipkart.poseidon.constants.RequestConstants;
 import com.flipkart.poseidon.core.PoseidonResponse;
-import com.flipkart.poseidon.handlers.http.utils.StringUtils;
 import com.flipkart.poseidon.internal.OrchestratorDataSource;
 import com.flipkart.poseidon.internal.ParamValidationFilter;
 import com.flipkart.poseidon.legoset.PoseidonLegoSet;
 import com.flipkart.poseidon.mappers.Mapper;
-import com.flipkart.poseidon.model.VariableModel;
 import com.flipkart.poseidon.pojos.EndpointPOJO;
 import com.flipkart.poseidon.pojos.ParamPOJO;
+import com.flipkart.poseidon.utils.ApiHelper;
 import flipkart.lego.api.entities.*;
 import flipkart.lego.api.exceptions.ElementNotFoundException;
 import flipkart.lego.api.exceptions.InternalErrorException;
@@ -138,42 +136,7 @@ public class APIBuildable implements Buildable {
         }
 
         for (ParamPOJO param : params) {
-            final JavaType javaType;
-            if (param.getType() != null) {
-                javaType = constructJavaType(param.getType());
-            } else if (!StringUtils.isNullOrEmpty(param.getJavatype())) {
-                javaType = constructJavaType(param.getJavatype());
-            } else {
-                javaType = null;
-            }
-            param.setJavaType(javaType);
+            param.setJavaType(ApiHelper.constructJavaType(param));
         }
-    }
-
-    private JavaType constructJavaType(String type) {
-        Class<?> clazz;
-        try {
-            clazz = Class.forName(type);
-        } catch (ClassNotFoundException e) {
-            throw new UnsupportedOperationException("Specify a known class " + type, e);
-        }
-
-        return configuration.getObjectMapper().getTypeFactory().constructType(clazz);
-    }
-
-    private JavaType constructJavaType(VariableModel variableModel) {
-        Class<?> clazz;
-        try {
-            clazz = Class.forName(variableModel.getType());
-        } catch (ClassNotFoundException e) {
-            throw new UnsupportedOperationException("Specify a known class " + variableModel.getType(), e);
-        }
-
-        JavaType[] javaTypes = new JavaType[variableModel.getTypes().length];
-        for (int i = 0; i < variableModel.getTypes().length; i++) {
-            javaTypes[i] = constructJavaType(variableModel.getTypes()[i]);
-        }
-
-        return configuration.getObjectMapper().getTypeFactory().constructParametrizedType(clazz, clazz, javaTypes);
     }
 }