+++ common/src/main/java/com/flipkart/poseidon/model/exception/MissingInformationException.java
@@ -4,4 +4,10 @@ package com.flipkart.poseidon.model.exception;
  * Created by shrey.garg on 01/11/16.
  */
 public class MissingInformationException extends Exception {
+    public MissingInformationException() {
+    }
+
+    public MissingInformationException(String message) {
+        super(message);
+    }
 }