+++ core/src/main/java/com/flipkart/poseidon/datasources/AbstractDataSource.java
@@ -54,6 +54,17 @@ public abstract class AbstractDataSource<T extends DataType> implements DataSour
         return dataSource.call();
     }
 
+    protected <Q extends DataType> Q executeSync(String dsId, Map<String, Object> requestMap) throws Exception {
+        DataSourceRequest dataSourceRequest = new DataSourceRequest();
+        dataSourceRequest.setAttributes(requestMap);
+        return executeSync(dsId, dataSourceRequest);
+    }
+
+    protected <Q extends DataType> Q executeSync(String dsId, Request request) throws Exception {
+        DataSource<Q> dataSource = this.legoset.getBasicDataSource(dsId, request);
+        return dataSource.call();
+    }
+
     @Override
     public abstract T call() throws Exception;