+++ sample/src/main/java/com/flipkart/poseidon/sample/datatypes/UserDataType.java
@@ -16,14 +16,8 @@
 
 package com.flipkart.poseidon.sample.datatypes;
 
-import com.flipkart.poseidon.model.annotations.Description;
-import com.flipkart.poseidon.model.annotations.Name;
-import com.flipkart.poseidon.model.annotations.Version;
 import flipkart.lego.api.entities.DataType;
 
-@Name("UserDataType")
-@Version(major = 1, minor = 0, patch = 0)
-@Description("User Entity")
 public class UserDataType implements DataType {
     private int userId;
     private String userName;