+++ service-clients-core/src/main/java/com/flipkart/poseidon/serviceclients/ServiceResponseDecoder.java
@@ -95,7 +95,18 @@ public class ServiceResponseDecoder<T> implements HttpResponseDecoder<ServiceRes
             } else {
                 try {
                     // Don't deserialize a plain string response using jackson
-                    final JavaType javaType = serviceResponseInfoMap.get("200").getType();
+                    final ServiceResponseInfo serviceResponseInfo = serviceResponseInfoMap.get(String.valueOf(statusCode));
+                    final JavaType javaType;
+                    if (serviceResponseInfo != null) {
+                        javaType = serviceResponseInfo.getType();
+                    } else {
+                        javaType = serviceResponseInfoMap.get("200").getType();
+                    }
+
+                    if (javaType == null) {
+                        return new ServiceResponse<T>((T) null, headers);
+                    }
+
                     if (String.class.isAssignableFrom(javaType.getRawClass())) {
                         return new ServiceResponse<T>((T) IOUtils.toString(httpResponse.getEntity().getContent()), headers);
                     }